Add rols

+++ back/src/controllers/users.ts
@@ -1,6 +1,7 @@
 import { Request, Response } from "express";
 import UserModel from "../models/user";
 import { hashPassword, comparePasswords, generateToken, verifyToken  } from '../middlewares';
+import jwt from "jsonwebtoken";
 
 
 
@@ -8,17 +9,18 @@ interface UserProps extends Omit<UserModel, "id" | "createdAt" | "updatedAt"> {}
 
 function formatedData (item: UserModel) {
   const user = {
-       id: item.id,
-       name: item.name,
-       surname: item.surname,
-       email: item.email,
-       age: item.age,
-       job: item.job,
-       isAdult: item.isAdult,
-       gender: item.gender,
-      //  createdAt: item.createdAt,
-      //  updatedAt: item.updatedAt
-     };
+    id: item.id,
+    name: item.name,
+    surname: item.surname,
+    email: item.email,
+    age: item.age,
+    job: item.job,
+    isAdult: item.isAdult,
+    gender: item.gender,
+    role: item.role,
+    //  createdAt: item.createdAt,
+    //  updatedAt: item.updatedAt
+  };
     return user
 };
 
@@ -59,6 +61,7 @@ export const addUser = async (req: Request, res: Response): Promise<void> => {
     const newUser = await UserModel.create({
       ...user,
       password: hashedPassword,
+      role: "user",
       isAdult,
     });
 
@@ -166,8 +169,25 @@ export const updateUser = async (req: Request, res: Response): Promise<void> => 
 export const deleteUser = async (req: Request, res: Response): Promise<void> => {
 
   const id = req.params.id;
+  const token = req.cookies[`my-token-${id}`];
 
   try {
+     if (!token) {
+       res.status(403).json({
+         message: "Forbidden: Token not provided",
+       });
+       return;
+     };
+
+     const decodedToken: any = jwt.verify(token, `${process.env.SECRET_KEY}`); 
+  
+     if (decodedToken.userId !== id) {
+       res.status(403).json({
+         message:
+           "Forbidden: You don't have permission to delete this user",
+       });
+       return;
+     };
     
     await UserModel.destroy({
       where: {
